+++ zqd/api/api.go
@@ -2,25 +2,12 @@ package api
 
 import (
 	"encoding/json"
-	"errors"
-	"net/http"
 	"net/url"
+	"strconv"
 
 	"github.com/brimsec/zq/pkg/nano"
 	"github.com/brimsec/zq/zio/zjsonio"
-	"github.com/gorilla/mux"
 )
-
-// ExtractSpace returns the unescaped space from the path of a request.
-// XXX is this in the right package?
-func ExtractSpace(r *http.Request) (string, error) {
-	v := mux.Vars(r)
-	space, ok := v["space"]
-	if !ok {
-		return "", errors.New("no space found")
-	}
-	return url.PathUnescape(space)
-}
 
 type Error struct {
 	Type    string      `json:"type"`
@@ -91,4 +78,83 @@ type SpaceInfo struct {
 type StatusResponse struct {
 	Ok      bool   `json:"ok"`
 	Version string `json:"version"`
+}
+
+// PacketSearch are the query string args to the packet endpoint when searching
+// for packets within a connection 5-tuple.
+type PacketSearch struct {
+	Span    nano.Span
+	Proto   string `validate:"required"`
+	SrcHost string `validate:"required"`
+	SrcPort *uint16
+	DstHost string `validate:"required"`
+	DstPort *uint16
+}
+
+// ToQuery transforms a packet search into a url.Values.
+func (ps *PacketSearch) ToQuery() url.Values {
+	tssec, tsns := ps.Span.Ts.Split()
+	dursec := int(ps.Span.Dur / 1000000000)
+	durns := int(int64(ps.Span.Dur) - int64(dursec)*1000000000)
+	q := url.Values{}
+	q.Add("ts_sec", strconv.Itoa(int(tssec)))
+	q.Add("ts_ns", strconv.Itoa(int(tsns)))
+	q.Add("duration_sec", strconv.Itoa(dursec))
+	q.Add("duration_ns", strconv.Itoa(durns))
+	q.Add("proto", ps.Proto)
+	q.Add("src_host", ps.SrcHost)
+	q.Add("dst_host", ps.DstHost)
+	if ps.SrcPort != nil {
+		q.Add("src_port", strconv.Itoa(int(*ps.SrcPort)))
+	}
+	if ps.DstPort != nil {
+		q.Add("dst_port", strconv.Itoa(int(*ps.DstPort)))
+	}
+
+	return q
+}
+
+// FromQuery parses a query string and populates the receiver's values.
+func (ps *PacketSearch) FromQuery(v url.Values) error {
+	var err error
+	var tsSec, tsNs, durSec, durNs int64
+	if tsSec, err = strconv.ParseInt(v.Get("ts_sec"), 10, 64); err != nil {
+		return err
+	}
+	if tsNs, err = strconv.ParseInt(v.Get("ts_ns"), 10, 64); err != nil {
+		return err
+	}
+	if durSec, err = strconv.ParseInt(v.Get("duration_sec"), 10, 64); err != nil {
+		return err
+	}
+	if durNs, err = strconv.ParseInt(v.Get("duration_ns"), 10, 64); err != nil {
+		return err
+	}
+	if v.Get("src_port") != "" {
+		p, err := strconv.ParseUint(v.Get("src_port"), 10, 16)
+		if err != nil {
+			return err
+		}
+		sp := uint16(p)
+		ps.SrcPort = &sp
+	}
+	if v.Get("dst_port") != "" {
+		p, err := strconv.ParseUint(v.Get("dst_port"), 10, 16)
+		if err != nil {
+			return err
+		}
+		sp := uint16(p)
+		ps.DstPort = &sp
+	}
+
+	span := nano.Span{
+		Ts:  nano.Unix(tsSec, tsNs),
+		Dur: nano.Duration(durSec, durNs),
+	}
+
+	ps.Span = span
+	ps.Proto = v.Get("proto")
+	ps.SrcHost = v.Get("src_host")
+	ps.DstHost = v.Get("dst_host")
+	return err
 }
