+++ zqd/space/space.go
@@ -1,8 +1,6 @@
 package space
 
 import (
-	"encoding/json"
-	"net/http"
 	"os"
 	"path/filepath"
 
@@ -11,15 +9,9 @@ import (
 	"github.com/brimsec/zq/zng/resolver"
 	"github.com/brimsec/zq/zqd/api"
 	"github.com/brimsec/zq/zqd/pcap"
-	"github.com/gorilla/mux"
 )
 
-func AddRoutes(router *mux.Router) {
-	router = router.UseEncodedPath() // Allow url encoded spaces
-	router.HandleFunc("/space/{space}/", HandleInfo).Methods("GET")
-}
-
-func spaceInfo(path string) (*api.SpaceInfo, error) {
+func Info(path string) (*api.SpaceInfo, error) {
 	bzngFile := filepath.Join(path, "all.bzng")
 	info, err := os.Stat(bzngFile)
 	if err != nil {
@@ -63,21 +55,4 @@ func spaceInfo(path string) (*api.SpaceInfo, error) {
 		s.MaxTime = &maxTs
 	}
 	return s, nil
-}
-
-func HandleInfo(w http.ResponseWriter, r *http.Request) {
-	space, err := api.ExtractSpace(r)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-	// XXX this is slow.  can easily cache result rather than scanning
-	// whole file each time.
-	info, err := spaceInfo(space)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-	w.Header().Add("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(info)
 }
